extract to a helper class
