+++ src/Query/IteratorBuilder.php
@@ -2,8 +2,6 @@
 
 namespace Statamic\Query;
 
-use Statamic\Support\Str;
-
 abstract class IteratorBuilder extends Builder
 {
     protected $randomize = false;
@@ -141,53 +139,7 @@ abstract class IteratorBuilder extends Builder
 
     protected function getFilterItemValue($item, $name)
     {
-        $nameExploded = explode('->', $name);
-        while (! empty($nameExploded)) {
-            $name = array_shift($nameExploded);
-            $item = $this->getFilterItemPartValue($item, $name);
-            if (is_null($item)) {
-                return;
-            }
-        }
-
-        return $item;
-    }
-
-    // any changes to this method should also be reflected in Statamic\Stache\Indexes\Value::getItemPartValue()
-    private function getFilterItemPartValue($item, $name)
-    {
-        $method = Str::camel($name);
-
-        if ($method === 'blueprint') {
-            return $item->blueprint()->handle();
-        }
-
-        if ($method === 'entriesCount') {
-            return $item->entriesCount();
-        }
-
-        // Don't want to use the authors() method, which would happen right after this.
-        if ($method === 'authors') {
-            return $item->value('authors');
-        }
-
-        if (is_array($item)) {
-            return $item[$name] ?? null;
-        }
-
-        if (is_scalar($item)) {
-            return null;
-        }
-
-        if (method_exists($item, $method)) {
-            return $item->{$method}();
-        }
-
-        if (method_exists($item, 'value')) {
-            return $item->value($name);
-        }
-
-        return $item->get($name);
+        return (new ResolveValue)($item, $name);
     }
 
     abstract protected function getBaseItems();
